# src/main/resources/public/api-docs.yaml
...
openapi: '3.0.0'
info:
  title: Copper Developers Test
  description: |
    These API's are the implementation of the development test for copper.co. They allow for the sycnronisation of a user (defined by the client-id/client-secret) to be stored in a local persistence layer. It also allows for the retrieval of both deposits and withdrawals (however it wasn't possible to implement withdrawals, as the external bitcoin address won't clear in time). It also has API's that allow for transfer between the main account and the sub-account
  version: "0.0.1-SNAPSHOT"
  contact:
    name: "Darrell Merryweather"
    email: "darrell.merryweather@hotmail.com"
security:
  - BasicAuth: []
  - BearerAuth: []
paths:
...
  /accounts/{id}:
    parameters:
      - name: id
        required: true
        in: path
        description: The ID of the Account
        schema:
          type: string
        example: 1234
    get:
      tags:
        - Accounts
      description: |
        Retrieves the specified Account including a list of its portfolios.
      responses:
        '200':
          description: The Account that was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: The Account doesn't exist
    delete:
      tags:
        - Accounts
      description: |
        Deletes a specific Account.
      responses:
        '204':
          description: Template has been deleted.
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Account:
      required:
        - id
      properties:
        id:
          type: string
        type:
          type: string
        username:
          type: string
        email:
          type: string
        portfolios:
          $ref: '#/components/schemas/Portfolios'
      example:
        code: 12345
        type: main
        username: some-username
        email: some-email@some-domain.com
        portfolios:
          btc:
            currency: "BTC"
            balance: 8.0
            available_withdrawal_funds: 8.0
            available_funds: 8.0
...
    Portfolios:
      additionalProperties:
        $ref: '#/components/schemas/Portfolio'
    Portfolio:
      properties:
        currency:
          type: string
        balance:
          type: number
        available_withdrawal_funds:
          type: number
        available_funds:
          type: number
      example:
        currency: "btc"
        balance: 8.0
        available_withdrawal_funds: 8.0
        available_funds: 8.0